Update server.R
Added several features to make MetaNetwork functional from a single command in any R console:
Line 2: added the infrastructure to support many different organisms in the Allez enrichment by adding a database containing the organism name and the necessary information for the database related argument variables in the fetchMart and enrichAllez functions.

Line 30-49. Added citation; moved the assignment of the path variable to the beginning of the code with the rest of the variable assignments

Line 29: Removed superfulous comment

Line 42: moved allowWGCNAThreads to within the command flow.

Line 60/81: Fixed grammar

Line 94-112: Added tryCatch wrapper around the merge module function. If the merge module function fails because cutHeight is too low, it will increase cutHeight by 0.1. It will do this twice and then fail with a message.

Line 213-227: addition of database containing animals made these lines unneccessary, so I completely restructured it.

Lines 269-272: added citation to the Enrich_shiny application.

Line 283: If you were running the GO enrichment application in a new folder, the output wouldn't be written. Added the infrastructure to make sure it gets written to the Results folder.